Refactor restore to new project by breaking into smaller components
